Replaced YoutubePlayer web impl to use https://www.npmjs.com/package/react-youtube
