+++ packages/core/src/components/YoutubePlayer/YoutubePlayerProps.ts
@@ -1,0 +1,9 @@
+import { StyleProp, ViewStyle } from "react-native";
+
+export interface YoutubePlayerProps {
+  videoId?: string;
+  playlist?: string;
+  mute?: boolean;
+  autoplay?: boolean;
+  style?: StyleProp<ViewStyle>;
+}
