+++ packages/core/src/components/YoutubePlayer/index.ts
@@ -1,0 +1,1 @@
+export { default as YoutubePlayer } from "./YotubePlayer";
